Bug fix: can not cast from long to int

+++ src/android/plugin/google/maps/MyPluginLayout.java
@@ -512,7 +512,7 @@ public class MyPluginLayout extends FrameLayout implements ViewTreeObserver.OnSc
         String domIDs[] = HTMLNodes.keySet().toArray(new String[HTMLNodes.size()]);
         Bundle domInfo = HTMLNodes.get(pluginMap.mapDivId);
         RectF htmlElementRect;
-        int mapDivDepth = domInfo.getInt("depth");
+        long mapDivDepth = domInfo.getLong("depth");
 
         for (String domId : domIDs) {
           if (pluginMap.mapDivId.equals(domId)) {
@@ -525,7 +525,7 @@ public class MyPluginLayout extends FrameLayout implements ViewTreeObserver.OnSc
           if (domInfo == null) {
             continue;
           }
-          if (domInfo.getInt("depth") <= mapDivDepth) {
+          if (domInfo.getLong("depth") <= mapDivDepth) {
             continue;
           }
